pyproject.toml
[project]
name = "rogomatic-llm"
version = "0.1.0"
description = "Rogomatic for LLMs - play Rogue with any LLM that supports structured output"
readme = "README.md"
authors = [
    { name = "iwhalen", email = "ianpatrickwhalen@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=1.63.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
rogomatic-llm = "rogomatic_llm.__main__:cli"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
    "ty>=0.0.19",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["rogomatic_llm"]
